diagram_generator/requirements.txt
# Core dependencies
pyyaml>=6.0
hypothesis>=6.0  # For property-based testing

# Testing dependencies
pytest>=7.0
pytest-cov>=4.0

# Code quality
black>=23.0
mypy>=1.0
